[*] Fix Piano fix [*] Comment gameMap

+++ src/game/audio/AudioPlayer.java
@@ -7,15 +7,14 @@ import java.util.List;
 
 public class AudioPlayer {
 
-	private List<Source> afx, noRemoveAFX, toRemove;
+	private List<Source> afx, toRemove;
 	private Source musicSource;
 
 	public AudioPlayer() {
 		afx = new ArrayList<>();
 		toRemove = new ArrayList<>();
-		this.noRemoveAFX = new ArrayList<>();
 
-		musicSource = new Source();
+		musicSource = new Source(false);
 		musicSource.setVolume(Options.musicVolume);
 		musicSource.setLooping(true);
 	}
@@ -24,32 +23,32 @@ public class AudioPlayer {
 		return musicSource;
 	}
 
-	public void playAfx(Sound afx) {
-		playAfx(new Source(0, 0, 0, Options.effectVolume), afx);
+	public void playAfx(Sound afx, boolean deleteOnFinish) {
+		playAfx(new Source(0, 0, 0, Options.effectVolume, deleteOnFinish), afx);
 	}
 
 	public void playAfx(Source c, Sound afx) {
-		this.noRemoveAFX.add(c);
+		this.afx.add(c);
 		c.play(afx);
 		c.setVolume(Options.effectVolume);
 	}
 
-	public void playAfx(Sound afx, float x, float y, float z) {
-		playAfx(new Source(x, y, z, Options.effectVolume), afx);
+	public void playAfx(Sound afx, float x, float y, float z, boolean deleteOnFinish) {
+		playAfx(new Source(x, y, z, Options.effectVolume, deleteOnFinish), afx);
 	}
 
-	public void playAfx(String afx) {
-		playAfx(new Source(0, 0, 0, Options.effectVolume), afx);
+	public void playAfx(String afx, boolean deleteOnFinish) {
+		playAfx(new Source(0, 0, 0, Options.effectVolume, deleteOnFinish), afx);
 	}
 
 	public void playAfx(Source c, String afx) {
-		this.noRemoveAFX.add(c);
+		this.afx.add(c);
 		c.play(afx);
 		c.setVolume(Options.effectVolume);
 	}
 
-	public void playAfx(String afx, float x, float y, float z) {
-		playAfx(new Source(x, y, z, Options.effectVolume), afx);
+	public void playAfx(String afx, float x, float y, float z, boolean deleteOnFinish) {
+		playAfx(new Source(x, y, z, Options.effectVolume, deleteOnFinish), afx);
 	}
 
 	public void update() {
@@ -65,7 +64,7 @@ public class AudioPlayer {
 			Source c = toRemove.get(i);
 			afx.remove(c);
 			c.stop();
-			c.delete();
+			if (c.shouldDeleteOnFinish()) c.delete();
 		}
 
 		toRemove.clear();
@@ -74,19 +73,5 @@ public class AudioPlayer {
 			Source c = afx.get(i);
 			c.setVolume(Options.effectVolume);
 		}
-		for (int i = 0; i < noRemoveAFX.size(); i++) {
-			Source c = noRemoveAFX.get(i);
-			c.setVolume(Options.effectVolume);
-		}
-	}
-
-	public boolean delete(Source c) {
-		if(noRemoveAFX.contains(c)) {
-			noRemoveAFX.remove(c);
-			c.stop();
-			c.delete();
-			return true;
-		}
-		return false;
 	}
 }
